tests/wpt/web-platform-tests/paint-timing/fcp-only/fcp-pseudo-element-image.html

<!DOCTYPE html>
<head>
<title>Performance Paint Timing Test: FCP due to pseudo-element image</title>
<style>
    #main {
        position: relative;
        width: 100px;
        height: 100px;
    }

    /* contentful class is defined in test_fcp script. */
    #main.contentful:after {
        content: url(../resources/circles.png);
    }
</style>
</head>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/utils.js"></script>
<div id="main">
</div>
<script>
    // Load the image into memory first to make sure it's decoded.
    function load_image() {
      const img = document.createElement("img");
      img.src = "../resources/circles.png";

      return new Promise(resolve => {
        img.onload = async () => resolve();
      });
    }

    test_fcp("First contentful paint fires due to pseudo-element image.", load_image)
</script>
</body>

</html>
